Refactor health and sync managers with new architecture and improved sync logic
